[Fix] 출력에 관한 책임 모두 console로 옮기기
