+++ src/main/java/com/programmers/calculator/engine/io/Output.java
@@ -1,7 +1,11 @@
 package com.programmers.calculator.engine.io;
 
-public interface Output {
-    void inputError(Exception e);
+import java.util.Map;
 
+public interface Output {
     void endNotification();
+
+    void calculateResult(String s);
+
+    void formulaList(Map<Integer, String> map);
 }
