+++ src/main/java/com/programmers/calculator/engine/storage/Storage.java
@@ -1,7 +1,9 @@
 package com.programmers.calculator.engine.storage;
 
+import java.util.Map;
+
 public interface Storage {
     void save(String s);
 
-    void findAll();
+    Map<Integer, String> findAll();
 }
